+++ vscode/test/e2e/at-file-context-selecting.test.ts
@@ -1,3 +1,5 @@
+import path from 'path'
+
 import { expect } from '@playwright/test'
 
 import { sidebarSignin } from './common'
@@ -30,6 +32,20 @@ test('@-file empty state', async ({ page, sidebar }) => {
     await chatInput.fill('@#')
     await expect(chatPanelFrame.getByRole('heading', { name: 'Search for a symbol to include..' })).toBeVisible()
 
+    // Forward slashes
+    await chatInput.fill('@lib/batches/env')
+    await expect(
+        chatPanelFrame.getByRole('button', { name: withPlatformSlashes('lib/batches/env/var.go') })
+    ).toBeVisible()
+
+    // Backslashes
+    if (path.sep === path.win32.sep) {
+        await chatInput.fill('@lib\\batches\\env')
+        await expect(
+            chatPanelFrame.getByRole('button', { name: withPlatformSlashes('lib/batches/env/var.go') })
+        ).toBeVisible()
+    }
+
     // Searching and clicking
     await chatInput.fill('Explain @mj')
     await chatPanelFrame.getByRole('button', { name: 'Main.java' }).click()
@@ -41,21 +57,23 @@ test('@-file empty state', async ({ page, sidebar }) => {
     // Keyboard nav
     await chatInput.type('Explain @vgo', { delay: 50 }) // without this delay the following Enter submits the form instead of selecting
     await chatInput.press('Enter')
-    await expect(chatInput).toHaveValue('Explain @lib/batches/env/var.go ')
+    await expect(chatInput).toHaveValue(withPlatformSlashes('Explain @lib/batches/env/var.go '))
     await chatInput.type('and @vgo', { delay: 50 }) // without this delay the following Enter submits the form instead of selecting
     await chatInput.press('ArrowDown') // second item
     await chatInput.press('ArrowDown') // wraps back to first item
     await chatInput.press('ArrowDown') // second item again
     await chatInput.press('Enter')
     await expect(chatInput).toHaveValue(
-        'Explain @lib/batches/env/var.go and @lib/codeintel/tools/lsif-visualize/visualize.go '
+        withPlatformSlashes('Explain @lib/batches/env/var.go and @lib/codeintel/tools/lsif-visualize/visualize.go ')
     )
 
     // Send the message and check it was included
     await chatInput.press('Enter')
     await expect(chatInput).toBeEmpty()
     await expect(
-        chatPanelFrame.getByText('Explain @lib/batches/env/var.go and @lib/codeintel/tools/lsif-visualize/visualize.go')
+        chatPanelFrame.getByText(
+            withPlatformSlashes('Explain @lib/batches/env/var.go and @lib/codeintel/tools/lsif-visualize/visualize.go')
+        )
     ).toBeVisible()
 
     // Check pressing tab after typing a complete filename.
@@ -66,3 +84,7 @@ test('@-file empty state', async ({ page, sidebar }) => {
     await chatInput.press('Tab')
     await expect(chatInput).toHaveValue('@Main.java ')
 })
+
+function withPlatformSlashes(input: string) {
+    return input.replaceAll(path.posix.sep, path.sep)
+}