+++ agent/src/agent.ts
@@ -400,7 +400,7 @@ export class Agent extends MessageHandler {
             if (!client) {
                 throw new Error('Cody client not initialized')
             }
-            const res = await client.graphqlClient.getCurrentUserIdAndVerifiedEmailAndCodyPro()
+            const res = await client.graphqlClient.getCurrentUserCodyProEnabled()
             if (res instanceof Error) {
                 throw res
             }