+++ vscode/CHANGELOG.md
@@ -14,6 +14,8 @@ This is a log of all notable changes to Cody for VS Code. [Unreleased] changes a
 
 ### Changed
 
+- Replace "Sign Out" with an account dialog. [pull/2233](https://github.com/sourcegraph/cody/pull/2233)
+
 ## [0.18.2]
 
 ### Added